# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Glassnode clone for use in Iran",
        "title": "Glassnode Service",
        "contact": {
            "name": "ArioDev",
            "url": "https://ariodev.com"
        },
        "version": "1"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/api/endpoint/all": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Endpoints"
                ],
                "summary": "Get the supported endpoints",
                "operationId": "endpoints_get",
                "responses": {
                    "200": {
                        "description": "Successfull",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.EndpointGetAll"
                            }
                        }
                    }
                }
            }
        },
        "/api/endpoint/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Endpoint"
                ],
                "summary": "Get the supported endpoint",
                "operationId": "endpoints_get_one",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Endpoint id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfull",
                        "schema": {
                            "$ref": "#/definitions/types.EndpointGetAll"
                        }
                    }
                }
            }
        },
        "/api/middlegroup": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups Management"
                ],
                "summary": "Create the middle group",
                "operationId": "middlegroup_create",
                "parameters": [
                    {
                        "description": "Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.MiddleGroupCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/types.CreateTopGroupRes"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/types.CreateTopGroupRes"
                        }
                    },
                    "500": {
                        "description": "Server faild",
                        "schema": {
                            "$ref": "#/definitions/types.CreateTopGroupRes"
                        }
                    }
                }
            }
        },
        "/api/middlegroup/{id}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups Management"
                ],
                "summary": "Delete the middle group",
                "operationId": "middlegroup_delete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted",
                        "schema": {
                            "$ref": "#/definitions/types.CreateTopGroupRes"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/types.CreateTopGroupRes"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups Management"
                ],
                "summary": "Update the middle group",
                "operationId": "middlegroup_update",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Group id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updating data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.MiddleGroupUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated",
                        "schema": {
                            "$ref": "#/definitions/types.CreateTopGroupRes"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/types.CreateTopGroupRes"
                        }
                    },
                    "500": {
                        "description": "Server faild",
                        "schema": {
                            "$ref": "#/definitions/types.CreateTopGroupRes"
                        }
                    }
                }
            }
        },
        "/api/topgroup": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups Management"
                ],
                "summary": "Get TopGroups",
                "operationId": "topgroups_get",
                "responses": {
                    "200": {
                        "description": "TopGroup",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.GetTopGroups"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups Management"
                ],
                "summary": "Add a new Top group",
                "operationId": "topgroup_create",
                "parameters": [
                    {
                        "description": "TopGroup name",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateTopGroupReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "TopGroup created",
                        "schema": {
                            "$ref": "#/definitions/types.CreateTopGroupRes"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/types.CreateTopGroupRes"
                        }
                    },
                    "500": {
                        "description": "Creating faild",
                        "schema": {
                            "$ref": "#/definitions/types.CreateTopGroupRes"
                        }
                    }
                }
            }
        },
        "/api/topgroup/{id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups Management"
                ],
                "summary": "Delete TopGroup",
                "operationId": "topgroups_delete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "TopGroup ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Result",
                        "schema": {
                            "$ref": "#/definitions/types.CreateTopGroupRes"
                        }
                    },
                    "500": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/types.CreateTopGroupRes"
                        }
                    }
                }
            },
            "patch": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups Management"
                ],
                "summary": "Update TopGroups",
                "operationId": "topgroups_update",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "TopGroup ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateTopGroupReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "TopGroup updated",
                        "schema": {
                            "$ref": "#/definitions/types.CreateTopGroupRes"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/types.CreateTopGroupRes"
                        }
                    },
                    "500": {
                        "description": "Updating faild",
                        "schema": {
                            "$ref": "#/definitions/types.CreateTopGroupRes"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "types.ChildGroups": {
            "type": "object",
            "properties": {
                "endpoint_id": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "types.CreateTopGroupReq": {
            "description": "Required data to create the Top Group",
            "type": "object",
            "properties": {
                "name": {
                    "description": "Name",
                    "type": "string"
                }
            }
        },
        "types.CreateTopGroupRes": {
            "description": "Top Group response data",
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "types.EndpointGetAll": {
            "type": "object",
            "properties": {
                "assets": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "currencies": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "formats": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                },
                "resolutions": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "tier": {
                    "type": "integer"
                }
            }
        },
        "types.GetTopGroups": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "middle_groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.MiddleGroup"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "types.MiddleGroup": {
            "type": "object",
            "properties": {
                "child_groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.ChildGroups"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "types.MiddleGroupCreate": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "topgroup_id": {
                    "type": "integer"
                }
            }
        },
        "types.MiddleGroupUpdate": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        }
    }
}